requirements.txt
aiokafka==0.10.0
fastapi==0.111.0
pandas==2.0.3
pydantic==1.10.8
python-dotenv==1.0.1
SQLAlchemy==2.0.27
redis==3.5.3
psycopg2-binary==2.9.9
psycopg2==2.9.9
openpyxl==3.0.7
uvicorn==0.30.1
streamlit==1.36.0